requirements.txt
# Python Dependencies cho Crypto Market Analysis Project
# Cài đặt: pip install -r requirements.txt

# HTTP requests và API clients
requests==2.31.0

# Database connectivity
mysql-connector-python==8.1.0

# Data processing và analysis
pandas==2.1.0
numpy==1.24.3

# Technical Analysis Library (tính toán chỉ báo kỹ thuật: RSI, MACD, Bollinger Bands, etc.)
ta==0.11.0

# Date và time utilities (thường được pandas sử dụng)
python-dateutil==2.8.2
